Display timeline of the patient journey (#18)
* Add TimelineJson to User Schema

TimelineJson is been added to the user Modal

* Add timeline Page

The timeline page has been added and the UI has been refined. Right now the demo text is used in the events array.

* Add FetchTimeline API

The api to fetch timeline is added. However, there is some authorization issue.

* fix: remove unnecessary assignment of timelineJson

* fix: use proper status code for Unauthorized

the earlier error was Content-Length required

* feat: allow Modal to be closed if clicking outside it

* feat: add TimelineModal to display timeline

* feat: assign timeline to the user

* fix: register /timeline endpoint on the middleware

* feat: show timeline modal upon click button

* fix: type errors

---------

Co-authored-by: Pankaj Singh <pankaj.s@launchxlabs.ai>
Co-authored-by: 1geek0 <nilay@getvitt.ai>