Add DemoMode Variable to the localstorage
Demo mode previously used a global context which set the toggle of whether the demo mode should be on or not. However on a refresh this toggle would set back to default value i.e false on. So rather than storing the demoMode toggle on global context, i stored it in the local sotrage.