Demo mode #11 (#14)
* feat: Prefilled Demo mode #11

* feat: fill in the text only upon clicking a button

* fix: import error

* fix: Error in submit prefill story

* Add Demo Stories.

add demo stories for demo mode. even added context. however encountering issue because demo mode cannot be opened without token.

* Add Demo Timeline Events data

Added demo timeline events data used for demo mode timeline modal page

* Update page.tsx

* Attached SignIn to Demo Mode

Demo Mode now requires sign in, which authenticated the /home and /notes route. Also since the demo mode is on, there would demo stories and timeline and there won't be any other API hits.

* Add DemoMode Variable to the localstorage

Demo mode previously used a global context which set the toggle of whether the demo mode should be on or not. However on a refresh this toggle would set back to default value i.e false on. So rather than storing the demoMode toggle on global context, i stored it in the local sotrage.

* Landing page (#27)

* first landing page

* hero changes

* Invalid Date issue resolved (#31)

The format of the date coming from the database and the format of the date on Story interface were different which caused the invalid date issue.

* Update TimelineModal.tsx (#25)

* Update TimelineModal.tsx

Removed the grey vertical bar and centered the calendar icon

* Update TimelineModal.tsx

UI Updates in timeline modal.

* Sync airtable resources to mongodb #10 (#21)

* base python script for #10

* Updated Mongo with Airtable resources.

* Update In Mongo Structure

* Updated Test Database

* Added delete functionality to prevent duplication.

---------

Co-authored-by: Ramshankar07 <picographer0214@gmail.com>

* Implement discourse for community #4 (#22)

* feat: #10 Airtable Import 'Patient Relevant Resources'

* Feat: Base-Setup-for-discourse-SSO

* fix: path error and logging information

* fix: disabling require_auth

* fix: del unwanted files

* fix: del airtable

* fix: redirect to discourse url

* fix: merge conflict

* fix: merge conflict

* fix: waiting for nonce

* Demo Mode merged with the latest main

---------

Co-authored-by: Pankaj Singh <pankaj.s@launchxlabs.ai>
Co-authored-by: Sainesh Nakra <30744887+saineshnakra@users.noreply.github.com>
Co-authored-by: Manas Goel <93397435+manasgoel2002@users.noreply.github.com>