60 who else button client side issue (#64)
* feat: Prefilled Demo mode #11

* feat: fill in the text only upon clicking a button

* fix: import error

* fix: Error in submit prefill story

* Add Demo Stories.

add demo stories for demo mode. even added context. however encountering issue because demo mode cannot be opened without token.

* Add Demo Timeline Events data

Added demo timeline events data used for demo mode timeline modal page

* Update page.tsx

* Attached SignIn to Demo Mode

Demo Mode now requires sign in, which authenticated the /home and /notes route. Also since the demo mode is on, there would demo stories and timeline and there won't be any other API hits.

* Add DemoMode Variable to the localstorage

Demo mode previously used a global context which set the toggle of whether the demo mode should be on or not. However on a refresh this toggle would set back to default value i.e false on. So rather than storing the demoMode toggle on global context, i stored it in the local sotrage.

* Provided unique keys to each element of map method

Earlier key was given the value of story.id. But not i have changed it to index of each element of the similarSotries array.

* New Template being passed in the prompt

* Add symptoms to the format.

* Adjust Date position.

* Update embeddings.ts

* Change Demo data for demo Mode

* Add sign in Demo Mode

* Disable Add Story and edit story buttons for Demo Mode

* Add Researches on the Demo Mode.

* Add public Resource to Demo Mode

* Chane Type of params

* Change Disclaimer on the timeline modal

When no stories are added or there is no time generated, the disclaimer on timeline modal now shows "Add more to your story so we can map your timeline together

* Add generateTitle function to the story api.

generateTitle was not being called in the story api and thus the title have been not been accurate.

* Fix Update Story Feature

The update story feature is now working perfectly. Needed to update the matcher in middlewares to fix this.

* Rename "Add New Story" button to "Update Story"

"Add New Story" Button on /notes has been changed to "Update Story"

* Null valued dates turned to 0

* Fixed the client side error for who else button.

---------

Co-authored-by: Ramshankar07 <picographer0214@gmail.com>